refactor(recorder): extract shared save and transcribe functionality
