+++ apps/app/src/lib/query/_queries/vadRecorder.ts
@@ -6,10 +6,7 @@ import type { WhisperingRecordingState } from '@repo/shared';
 import { nanoid } from 'nanoid/non-secure';
 import { defineMutation, defineQuery } from '../_utils';
 import { queryClient } from '../index';
-import { recordings } from './recordings';
-import { maybeCopyAndPaste } from '../../maybeCopyAndPaste';
-import { transcription } from './transcription';
-import { transformer } from './transformer';
+import { saveRecordingAndTranscribe } from '../../_handleRecordingTranscription';
 
 const vadRecorderKeys = {
 	all: ['vadRecorder'] as const,
@@ -55,134 +52,12 @@ export const vadRecorder = {
 					console.info('Voice activated speech captured');
 					services.sound.playSoundIfEnabled('vad-capture');
 
-					const now = new Date().toISOString();
-					const newRecordingId = nanoid();
-
-					const { data: createdRecording, error: createRecordingError } =
-						await recordings.createRecording.execute({
-							id: newRecordingId,
-							title: '',
-							subtitle: '',
-							createdAt: now,
-							updatedAt: now,
-							timestamp: now,
-							transcribedText: '',
-							blob,
-							transcriptionStatus: 'UNPROCESSED',
-						});
-					if (createRecordingError) {
-						toast.error({
-							id: toastId,
-							title: '❌ Database Save Failed',
-							description:
-								'Your voice activated capture was captured but could not be saved to the database. Please check your storage space and permissions.',
-							action: { type: 'more-details', error: createRecordingError },
-						});
-						return;
-					}
-					toast.loading({
-						id: toastId,
-						title: '✨ Voice activated capture complete!',
-						description: settings.value['recording.isFasterRerecordEnabled']
-							? 'Voice activated capture complete! Ready for another take'
-							: 'Voice activated capture complete! Session closed successfully',
-					});
-
-					const transcribeToastId = nanoid();
-					toast.loading({
-						id: transcribeToastId,
-						title: '📋 Transcribing...',
-						description: 'Your recording is being transcribed...',
-					});
-					const { data: transcribedText, error: transcribeError } =
-						await transcription.transcribeRecording.execute(createdRecording);
-
-					if (transcribeError) {
-						if (transcribeError.name === 'WhisperingError') {
-							toast.error({ id: transcribeToastId, ...transcribeError });
-							return;
-						}
-						toast.error({
-							id: transcribeToastId,
-							title: '❌ Failed to transcribe recording',
-							description: 'Your recording could not be transcribed.',
-							action: { type: 'more-details', error: transcribeError },
-						});
-						return;
-					}
-
-					toast.success({
-						id: transcribeToastId,
-						title: 'Transcribed recording!',
-						description: 'Your recording has been transcribed.',
-					});
-					maybeCopyAndPaste({
-						text: transcribedText,
+					await saveRecordingAndTranscribe({
+						blob,
 						toastId,
-						shouldCopy: settings.value['transcription.clipboard.copyOnSuccess'],
-						shouldPaste:
-							settings.value['transcription.clipboard.pasteOnSuccess'],
-						statusToToastText(status) {
-							switch (status) {
-								case null:
-									return '📝 Recording transcribed!';
-								case 'COPIED':
-									return '📝 Recording transcribed and copied to clipboard!';
-								case 'COPIED+PASTED':
-									return '📝📋✍️ Recording transcribed, copied to clipboard, and pasted!';
-							}
-						},
-					});
-
-					// Do I need to transform this text further?
-					const needsTransformation = settings.value['transformations.selectedTransformationId'];
-					if (!needsTransformation) {
-						return; // We're done - no transformation needed
-					}
-
-					// Can I get the transformation I need?
-					const { data: transformation, error: getTransformationError } =
-						await services.db.getTransformationById(needsTransformation);
-
-					const couldNotRetrieveTransformation = getTransformationError;
-					const transformationNoLongerExists = !transformation;
-
-					if (couldNotRetrieveTransformation) {
-						toast.error({
-							id: nanoid(),
-							title: '❌ Failed to get transformation',
-							description:
-								'Your transformation could not be retrieved. Please try again.',
-							action: { type: 'more-details', error: getTransformationError },
-						});
-						return;
-					}
-
-					if (transformationNoLongerExists) {
-						settings.value = {
-							...settings.value,
-							'transformations.selectedTransformationId': null,
-						};
-						toast.warning({
-							id: nanoid(),
-							title: '⚠️ No matching transformation found',
-							description:
-								'No matching transformation found. Please select a different transformation.',
-							action: {
-								type: 'link',
-								label: 'Select a different transformation',
-								goto: '/transformations',
-							},
-						});
-						return;
-					}
-
-					// Apply the transformation
-					const transformToastId = nanoid();
-					await transformer.transformRecording.execute({
-						recordingId: createdRecording.id,
-						transformationId: needsTransformation,
-						toastId: transformToastId,
+						completionTitle: '✨ Voice activated capture complete!',
+						completionDescription:
+							'Voice activated capture complete! Ready for another take',
 					});
 				},
 			});
